feat: improved navigation and added sponsor link
