+++ pkg/dashboard/frontend/src/components/layout/AppLayout/NavigationItem.tsx
@@ -1,0 +1,57 @@
+import { cn } from '@/lib/utils'
+import React from 'react'
+
+export interface NavigationItemProps {
+  name: string
+  href: string
+  icon: React.ForwardRefExoticComponent<
+    Omit<React.SVGProps<SVGSVGElement>, 'ref'> & {
+      title?: string
+      titleId?: string
+    } & React.RefAttributes<SVGSVGElement>
+  >
+  onClick?: () => void
+  routePath: string
+}
+
+const NavigationItem: React.FC<NavigationItemProps> = ({
+  name,
+  href,
+  icon: Icon,
+  onClick,
+  routePath,
+}) => {
+  const isActive = href === routePath
+
+  return (
+    <li key={name}>
+      <a
+        href={href}
+        onClick={onClick}
+        aria-current={isActive}
+        target={href.startsWith('http') ? '_blank' : undefined}
+        rel={href.startsWith('http') ? 'noopener noreferrer' : undefined}
+        className={cn(
+          isActive
+            ? 'bg-gray-100 text-primary shadow-sm'
+            : 'text-gray-400 hover:bg-gray-100 hover:text-primary',
+          'group relative flex h-12 w-12 items-center gap-x-3 rounded-md p-3 text-sm font-semibold leading-6 transition-all group-data-[state=expanded]:w-full',
+        )}
+      >
+        <Icon className="h-6 w-6 shrink-0" aria-hidden="true" />
+        <span
+          className={cn(
+            'min-w-[120px] text-sm',
+            'absolute left-7 group-data-[state=expanded]:left-12',
+            'opacity-0 group-data-[state=expanded]:opacity-100',
+            'transition-all',
+          )}
+        >
+          {name}
+        </span>
+      </a>
+    </li>
+  )
+}
+
+export default NavigationItem
