+++ pkg/dashboard/frontend/src/components/layout/AppLayout/index.ts
@@ -1,0 +1,3 @@
+import AppLayout from './AppLayout'
+
+export default AppLayout
