+++ pkg/dashboard/frontend/src/pages/architecture.astro
@@ -4,5 +4,5 @@ import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="Architecture | Local Dashboard | Nitric">
-  <ArchitectureExplorer client:only />
+  <ArchitectureExplorer client:only="react"="react" />
 </Layout>
