+++ pkg/dashboard/frontend/src/pages/index.astro
@@ -4,5 +4,5 @@ import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="APIs | Local Dashboard | Nitric">
-  <APIExplorer client:only />
+  <APIExplorer client:only="react" />
 </Layout>
