+++ pkg/dashboard/frontend/src/pages/schedules.astro
@@ -4,5 +4,5 @@ import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="Schedules | Local Dashboard | Nitric">
-  <EventExplorer client:only workerType="schedules" />
+  <EventExplorer client:only="react" workerType="schedules" />
 </Layout>
