+++ pkg/dashboard/frontend/src/pages/storage.astro
@@ -1,9 +1,8 @@
 ---
-import AppLayout from "../components/layout/AppLayout";
 import StorageExplorer from "../components/storage/StorageExplorer";
 import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="Storage | Local Dashboard | Nitric">
-  <StorageExplorer client:only />
+  <StorageExplorer client:only="react" />
 </Layout>
