+++ pkg/dashboard/frontend/src/pages/topics.astro
@@ -4,5 +4,5 @@ import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="Topics | Local Dashboard | Nitric">
-  <EventExplorer client:only workerType="topics" />
+  <EventExplorer client:only="react" workerType="topics" />
 </Layout>
