+++ pkg/dashboard/frontend/src/pages/websockets.astro
@@ -4,5 +4,5 @@ import Layout from "../layouts/Layout.astro";
 ---
 
 <Layout title="Websockets | Local Dashboard | Nitric">
-  <WSExplorer client:only />
+  <WSExplorer client:only="react" />
 </Layout>
